requirements.txt
fastapi
mangum
uvicorn
openai
psycopg2-binary
python-dotenv 